requirements.txt
# ── Core ──────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0

# ── Telegram ──────────────────────────────────────────────
python-telegram-bot>=20.7

# ── Database ──────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0

# ── ML ────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.2.0

# ── NLP / token estimation ────────────────────────────────
tiktoken>=0.6.0

# ── Utilities ─────────────────────────────────────────────
filelock>=3.13.0
tenacity>=8.2.0
rich>=13.7.0

# ── Dev / testing ─────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
